+++ internal/provider/resource_iosxr_gnmi.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-framework/types"
@@ -12,7 +11,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/netascode/terraform-provider-iosxr/internal/provider/client"
 	"github.com/netascode/terraform-provider-iosxr/internal/provider/helpers"
-	"github.com/tidwall/gjson"
 )
 
 type resourceGnmiType struct{}
@@ -59,6 +57,28 @@ func (t resourceGnmiType) GetSchema(ctx context.Context) (tfsdk.Schema, diag.Dia
 				Optional:            true,
 				Computed:            true,
 			},
+			"lists": {
+				MarkdownDescription: "YANG lists.",
+				Optional:            true,
+				Attributes: tfsdk.ListNestedAttributes(map[string]tfsdk.Attribute{
+					"name": {
+						MarkdownDescription: "YANG list name.",
+						Type:                types.StringType,
+						Required:            true,
+					},
+					"key": {
+						MarkdownDescription: "YANG list key attribute(s). In case of multiple keys, those should be separated by a comma (`,`).",
+						Type:                types.StringType,
+						Required:            true,
+					},
+					"items": {
+						Type:                types.ListType{ElemType: types.MapType{ElemType: types.StringType}},
+						MarkdownDescription: "List of maps of key-value pairs which represents the attributes and its values.",
+						Optional:            true,
+						Computed:            true,
+					},
+				}, tfsdk.ListNestedAttributesOptions{}),
+			},
 		},
 	}, nil
 }
@@ -85,9 +105,9 @@ func (r resourceGnmi) Create(ctx context.Context, req tfsdk.CreateResourceReques
 		return
 	}
 
-	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Create", plan.Id.Value))
+	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Create", plan.getPath()))
 
-	if !plan.Attributes.Unknown {
+	if !plan.Attributes.Unknown || len(plan.Lists) > 0 {
 		body := plan.toBody(ctx)
 
 		_, diags = r.provider.client.Set(ctx, plan.Device.Value, plan.Path.Value, body, client.Update)
@@ -124,25 +144,7 @@ func (r resourceGnmi) Read(ctx context.Context, req tfsdk.ReadResourceRequest, r
 		return
 	}
 
-	var ea map[string]string
-	state.Attributes.ElementsAs(ctx, &ea, false)
-	existingAttr := make([]string, len(ea))
-	for k := range ea {
-		existingAttr = append(existingAttr, k)
-	}
-
-	attributes := make(map[string]attr.Value)
-
-	for attr, value := range gjson.ParseBytes(getResp.Notification[0].Update[0].Val.GetJsonIetfVal()).Map() {
-		if helpers.Contains(existingAttr, attr) {
-			attributes[attr] = types.String{Value: value.String()}
-		}
-		// if len(existingAttr) > 0 && !helpers.Contains(existingAttr, attr) {
-		// 	continue
-		// }
-		// attributes[attr] = types.String{Value: value.String()}
-	}
-	state.Attributes.Elems = attributes
+	state.fromBody(ctx, getResp.Notification[0].Update[0].Val.GetJsonIetfVal())
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Read finished successfully", state.Id.Value))
 
@@ -151,10 +153,17 @@ func (r resourceGnmi) Read(ctx context.Context, req tfsdk.ReadResourceRequest, r
 }
 
 func (r resourceGnmi) Update(ctx context.Context, req tfsdk.UpdateResourceRequest, resp *tfsdk.UpdateResourceResponse) {
-	var plan Gnmi
+	var plan, state Gnmi
 
 	// Read plan
 	diags := req.Plan.Get(ctx, &plan)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	// Read state
+	diags = req.State.Get(ctx, &state)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
@@ -166,6 +175,17 @@ func (r resourceGnmi) Update(ctx context.Context, req tfsdk.UpdateResourceReques
 		body := plan.toBody(ctx)
 
 		_, diags = r.provider.client.Set(ctx, plan.Device.Value, plan.Path.Value, body, client.Update)
+		resp.Diagnostics.Append(diags...)
+		if resp.Diagnostics.HasError() {
+			return
+		}
+	}
+
+	deletedListItems := plan.getDeletedListItems(ctx, state)
+	tflog.Debug(ctx, fmt.Sprintf("List items to delete: %+v", deletedListItems))
+
+	for _, i := range deletedListItems {
+		_, diags := r.provider.client.Set(ctx, state.Device.Value, i, "", client.Delete)
 		resp.Diagnostics.Append(diags...)
 		if resp.Diagnostics.HasError() {
 			return