+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Add delete attribute to iosxr_gnmi resource
 - Allow provider config without host attribute in case devices attribute is being used
+- Enhance iosxr_gnmi resource to support nested attributes (within YANG containers) using "/" as separator
+- Enhance iosxr_gnmi resource to support nested YANG lists
 
 ## 0.1.0